qca-legacy-uboot: fix rootfs sign verify failed
If the bad block in rootfs partion, security sign verification
will fail, the root cause is nand_read() will skip the bad block
and read from the next block.
Add nand_read_block() to read only one block to fix the issue,
which will return -1 if the block is invalid.

Change-Id: I823de205069861656c78c281d135899be79b6f3b
Signed-off-by: Miaoqing Pan <miaoqing@codeaurora.org>